Additional Functionality - Mainly for testing
Login has been changed so that is not mandatory and is instead launched by a button.

Register button has been created which sends data from a dialog form to the controller which can then be posted to the server.

Added getData and postData buttons which are connected to two test methods within the controller. This was to ensure communication could be made between the server and client. GET has been tested however POST has not yet at this stage.