Added a richer user interface with connected controller operations
Have added a richer GUI with both a login or register main screens which
then takes the user to the MainScreen.

Developed a mainScreen to show the user there current books and allow
them to browse books from the server using JTables.

Have added in POST request methods within the controller so that login,
 register and new book data can be sent to the server.