Fixed the Gradle build to work with IntelliJ Designer
Had huge trouble using IntelliJ's in built layout designer with Gradle.
After researching have found out that the two just don't go together and
since it is not being updated any further there is no way to get it to
work on older versions.

Have added a fix but has ruined most of the layout.