Fixed Password Hasher and POST Request
Fixed an issue with salts generating different hashes each time. For now we have decided not to use salts and just hash passwords. This will be looked into at a further point in time.

Have added 2 new methods to controller which allow the retrieval of just the status of a POST request as well as getting data from a POST request.

Login, Register, User Books, All books, and load profile all now work successfully with the server.